refactor(config): extract newtype macro, debug_secret helper, and test separation
- Introduce define_lowercase_id! declarative macro to eliminate 4 identical newtype boilerplates (ProviderId, ChannelName, AgentId, BotId) from ~108 lines to ~30 lines
- Add debug_secret() helper to replace 5 duplicated redacted Debug impl patterns
- Move 1900-line test module from mod.rs to config/tests.rs (Rust convention)
- mod.rs reduced from 2325 lines to ~370 lines of production code

Ultraworked with [Sisyphus](https://github.com/code-yeongyu/oh-my-openagent)
